requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
SQLAlchemy==2.0.43
psycopg[binary]==3.2.9
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
httplib2==0.22.0
